refactor: localize live component styles
