+++ src/pages/Live/components/Live.module.css
@@ -1,0 +1,40 @@
+.section {
+    margin-bottom: 2.5rem;
+}
+
+.sectionBody {
+    border: 0.0625rem solid #ddd;
+    border-top: none;
+    padding: 1.25rem;
+    border-radius: 0 0 0.25rem 0.25rem;
+}
+
+.chartFilterRow {
+    margin-bottom: 0.3125rem;
+}
+
+.filterLabel {
+    margin-left: 0.625rem;
+}
+
+.intervalSelect {
+    margin-left: 0.625rem;
+}
+
+.deviceLabel {
+    text-align: center;
+    font-weight: bold;
+    margin-bottom: 0.625rem;
+}
+
+.spectrumBarChartWrapper {
+    width: 100%;
+    margin-top: 1.25rem;
+}
+
+@media (min-width: 48rem) {
+    .spectrumBarChartWrapper {
+        width: 80%;
+        margin: 1.25rem auto 0;
+    }
+}
