+++ src/pages/Live/components/Overview.module.css
@@ -4,49 +4,10 @@
     gap: 0.75rem;
 }
 
-.box {
-    background: #1f2937; /* slate-800 */
-    border-radius: 0.875rem;
-    padding: 0.875rem 1rem;
-    box-shadow: 0 0.125rem 0.625rem rgba(0,0,0,0.25);
-    color: #e5e7eb; /* gray-200 */
-    user-select: none;
-}
-
-.boxHeader { display: flex; align-items: center; justify-content: space-between; }
-.icon { font-size: 1.375rem; opacity: 0.9; }
-
-.valueRow { display: flex; align-items: baseline; gap: 0.375rem; margin-top: 0.375rem; }
-.value { font-size: 1.75rem; font-weight: 700; letter-spacing: 0.0125rem; }
-.unit { font-size: 0.875rem; opacity: 0.9; }
-
-.caption { margin-top: 0.5rem; }
-.title { font-size: 0.75rem; opacity: 0.9; }
-.subtitle { font-size: 0.6875rem; opacity: 0.7; margin-top: 0.125rem; }
-
-/* State colors */
-.ok { background: linear-gradient(180deg, #065f46 0%, #064e3b 100%); }      /* emerald */
-.warn { background: #fff3cd; color: #000; }    /* light yellow */
-.danger { background: #f8d7da; color: #000; }  /* light red */
-.neutral { background: linear-gradient(180deg, #374151 0%, #1f2937 100%); } /* slate */
-
 @media (max-width: 3rem) {
     .grid {
         grid-auto-flow: column;
         grid-auto-columns: minmax(7.5rem, 1fr);
         overflow-x: auto;
     }
-
-    .box {
-        padding: 0.625rem 0.75rem;
-    }
-
-    .icon { font-size: 1.25rem; }
-
-    .valueRow { margin-top: 0.25rem; }
-    .value { font-size: 1.375rem; }
-    .unit { font-size: 0.75rem; }
-
-    .title { font-size: 0.6875rem; }
-    .subtitle { font-size: 0.625rem; }
 }
