+++ src/pages/Live/components/SensorDashboard/SensorDashboard.module.css
@@ -1,0 +1,3 @@
+.dashboard {
+    padding: 1.25rem;
+}
