+++ src/pages/Live/components/SpectrumBarChart.module.css
@@ -1,0 +1,3 @@
+.chart {
+    width: 100%;
+}
