+++ src/pages/Live/components/TopicSection.module.css
@@ -1,0 +1,10 @@
+.deviceGroup {
+    margin-bottom: 2rem;
+}
+
+.topicTitle {
+    font-size: 1.2rem;
+    font-weight: bold;
+    margin-bottom: 0.75rem;
+    color: #333;
+}
